app/requirements.txt
Flask
Flask-Bcrypt
Flask-SQLAlchemy
Flask-Marshmallow
PyMySQL
PyJWT
pytest
marshmallow
python-dotenv
cryptography
sympy